refactor: split handler module
