+++ internal/handler/handler.go
@@ -1,23 +1,13 @@
 package handler
 
 import (
-	"context"
-	"fmt"
-	"github.com/go-telegram/bot"
-	"github.com/go-telegram/bot/models"
-	"log/slog"
 	"remnawave-tg-shop-bot/internal/cache"
-	"remnawave-tg-shop-bot/internal/config"
 	"remnawave-tg-shop-bot/internal/cryptopay"
 	"remnawave-tg-shop-bot/internal/database"
 	"remnawave-tg-shop-bot/internal/payment"
 	"remnawave-tg-shop-bot/internal/sync"
 	"remnawave-tg-shop-bot/internal/translation"
 	"remnawave-tg-shop-bot/internal/yookasa"
-	"remnawave-tg-shop-bot/utils"
-	"strconv"
-	"strings"
-	"time"
 )
 
 type Handler struct {
@@ -51,664 +41,4 @@ func NewHandler(
 		referralRepository: referralRepository,
 		cache:              cache,
 	}
-}
-
-func (h Handler) ReferralCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	customer, _ := h.customerRepository.FindByTelegramId(ctx, update.CallbackQuery.From.ID)
-	langCode := update.CallbackQuery.From.LanguageCode
-	refCode := customer.TelegramID
-
-	refLink := fmt.Sprintf("https://telegram.me/share/url?url=https://t.me/%s?start=ref_%d", update.CallbackQuery.Message.Message.From.Username, refCode)
-	count, err := h.referralRepository.CountByReferrer(ctx, customer.TelegramID)
-	if err != nil {
-		slog.Error("error counting referrals", err)
-		return
-	}
-	text := fmt.Sprintf(h.translation.GetText(langCode, "referral_text"), count)
-	callbackMessage := update.CallbackQuery.Message.Message
-	_, err = b.EditMessageText(ctx, &bot.EditMessageTextParams{
-		ChatID:    callbackMessage.Chat.ID,
-		MessageID: callbackMessage.ID,
-		Text:      text,
-		ParseMode: models.ParseModeHTML,
-		ReplyMarkup: models.InlineKeyboardMarkup{InlineKeyboard: [][]models.InlineKeyboardButton{
-			{
-				{Text: h.translation.GetText(langCode, "share_referral_button"), URL: refLink},
-			},
-			{
-				{Text: h.translation.GetText(langCode, "back_button"), CallbackData: CallbackStart},
-			},
-		}},
-	})
-	if err != nil {
-		slog.Error("Error sending referral message", err)
-	}
-}
-
-func (h Handler) StartCommandHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	ctxWithTime, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-	langCode := update.Message.From.LanguageCode
-	existingCustomer, err := h.customerRepository.FindByTelegramId(ctx, update.Message.Chat.ID)
-	if err != nil {
-		slog.Error("error finding customer by telegram id", err)
-		return
-	}
-
-	if existingCustomer == nil {
-		existingCustomer, err = h.customerRepository.Create(ctxWithTime, &database.Customer{
-			TelegramID: update.Message.Chat.ID,
-			Language:   langCode,
-		})
-		if err != nil {
-			slog.Error("error creating customer", err)
-			return
-		}
-
-		if strings.Contains(update.Message.Text, "ref_") {
-			arg := strings.Split(update.Message.Text, " ")[1]
-			if strings.HasPrefix(arg, "ref_") {
-				code := strings.TrimPrefix(arg, "ref_")
-				referrerId, err := strconv.ParseInt(code, 10, 64)
-				if err != nil {
-					slog.Error("error parsing referrer id", err)
-					return
-				}
-				_, err = h.customerRepository.FindByTelegramId(ctx, referrerId)
-				if err == nil {
-					_, err := h.referralRepository.Create(ctx, referrerId, existingCustomer.TelegramID)
-					if err != nil {
-						slog.Error("error creating referral", err)
-						return
-					}
-					slog.Info("referral created", "referrerId", utils.MaskHalfInt64(referrerId), "refereeId", utils.MaskHalfInt64(existingCustomer.TelegramID))
-				}
-			}
-		}
-	} else {
-		updates := map[string]interface{}{
-			"language": langCode,
-		}
-
-		err = h.customerRepository.UpdateFields(ctx, existingCustomer.ID, updates)
-		if err != nil {
-			slog.Error("Error updating customer", err)
-			return
-		}
-	}
-
-	inlineKeyboard := h.buildStartKeyboard(existingCustomer, langCode)
-
-	m, err := b.SendMessage(ctx, &bot.SendMessageParams{
-		ChatID: update.Message.Chat.ID,
-		Text:   "🧹",
-		ReplyMarkup: models.ReplyKeyboardRemove{
-			RemoveKeyboard: true,
-		},
-	})
-
-	if err != nil {
-		slog.Error("Error sending removing reply keyboard", err)
-		return
-	}
-
-	_, err = b.DeleteMessage(ctx, &bot.DeleteMessageParams{
-		ChatID:    update.Message.Chat.ID,
-		MessageID: m.ID,
-	})
-
-	if err != nil {
-		slog.Error("Error deleting message", err)
-		return
-	}
-
-	_, err = b.SendMessage(ctx, &bot.SendMessageParams{
-		ChatID:    update.Message.Chat.ID,
-		ParseMode: models.ParseModeHTML,
-		ReplyMarkup: models.InlineKeyboardMarkup{
-			InlineKeyboard: inlineKeyboard,
-		},
-		Text: h.translation.GetText(langCode, "greeting"),
-	})
-	if err != nil {
-		slog.Error("Error sending /start message", err)
-	}
-}
-
-func (h Handler) StartCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	ctxWithTime, cancel := context.WithTimeout(ctx, 5*time.Second)
-	defer cancel()
-
-	callback := update.CallbackQuery
-	langCode := callback.From.LanguageCode
-
-	existingCustomer, err := h.customerRepository.FindByTelegramId(ctxWithTime, callback.From.ID)
-	if err != nil {
-		slog.Error("error finding customer by telegram id", err)
-		return
-	}
-
-	inlineKeyboard := h.buildStartKeyboard(existingCustomer, langCode)
-
-	_, err = b.EditMessageText(ctxWithTime, &bot.EditMessageTextParams{ChatID: callback.Message.Message.Chat.ID,
-		MessageID: callback.Message.Message.ID,
-		ParseMode: models.ParseModeHTML,
-		ReplyMarkup: models.InlineKeyboardMarkup{
-			InlineKeyboard: inlineKeyboard,
-		},
-		Text: h.translation.GetText(langCode, "greeting"),
-	})
-	if err != nil {
-		slog.Error("Error sending /start message", err)
-	}
-}
-
-func (h Handler) CreateCustomerIfNotExistMiddleware(next bot.HandlerFunc) bot.HandlerFunc {
-	return func(ctx context.Context, b *bot.Bot, update *models.Update) {
-		var telegramId int64
-		var langCode string
-		if update.Message != nil {
-			telegramId = update.Message.From.ID
-			langCode = update.Message.From.LanguageCode
-		} else if update.CallbackQuery != nil {
-			telegramId = update.CallbackQuery.From.ID
-			langCode = update.CallbackQuery.From.LanguageCode
-		}
-		existingCustomer, err := h.customerRepository.FindByTelegramId(ctx, telegramId)
-		if err != nil {
-			slog.Error("error finding customer by telegram id", err)
-			return
-		}
-
-		if existingCustomer == nil {
-			existingCustomer, err = h.customerRepository.Create(ctx, &database.Customer{
-				TelegramID: telegramId,
-				Language:   langCode,
-			})
-			if err != nil {
-				slog.Error("error creating customer", err)
-				return
-			}
-		} else {
-			updates := map[string]interface{}{
-				"language": langCode,
-			}
-
-			err = h.customerRepository.UpdateFields(ctx, existingCustomer.ID, updates)
-			if err != nil {
-				slog.Error("Error updating customer", err)
-				return
-			}
-		}
-
-		next(ctx, b, update)
-	}
-}
-
-func (h Handler) resolveConnectButton(lang string) []models.InlineKeyboardButton {
-
-	var inlineKeyboard []models.InlineKeyboardButton
-
-	if config.GetMiniAppURL() != "" {
-		inlineKeyboard = []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(lang, "connect_button"), WebApp: &models.WebAppInfo{
-				URL: config.GetMiniAppURL(),
-			}},
-		}
-	} else {
-		inlineKeyboard = []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(lang, "connect_button"), CallbackData: CallbackConnect},
-		}
-	}
-	return inlineKeyboard
-}
-
-func (h Handler) buildStartKeyboard(existingCustomer *database.Customer, langCode string) [][]models.InlineKeyboardButton {
-	var inlineKeyboard [][]models.InlineKeyboardButton
-
-	if existingCustomer.SubscriptionLink == nil && config.TrialDays() > 0 {
-		inlineKeyboard = append(inlineKeyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "trial_button"), CallbackData: CallbackTrial},
-		})
-	}
-
-	inlineKeyboard = append(inlineKeyboard, [][]models.InlineKeyboardButton{
-		{{Text: h.translation.GetText(langCode, "buy_button"), CallbackData: CallbackBuy}},
-	}...)
-
-	if existingCustomer.SubscriptionLink != nil && existingCustomer.ExpireAt.After(time.Now()) {
-		inlineKeyboard = append(inlineKeyboard, h.resolveConnectButton(langCode))
-	}
-
-	if config.GetReferralDays() > 0 {
-		inlineKeyboard = append(inlineKeyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "referral_button"), CallbackData: CallbackReferral},
-		})
-	}
-
-	if config.ServerStatusURL() != "" {
-		inlineKeyboard = append(inlineKeyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "server_status_button"), URL: config.ServerStatusURL()},
-		})
-	}
-
-	if config.SupportURL() != "" {
-		inlineKeyboard = append(inlineKeyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "support_button"), URL: config.SupportURL()},
-		})
-	}
-
-	if config.FeedbackURL() != "" {
-		inlineKeyboard = append(inlineKeyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "feedback_button"), URL: config.FeedbackURL()},
-		})
-	}
-
-	if config.ChannelURL() != "" {
-		inlineKeyboard = append(inlineKeyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "channel_button"), URL: config.ChannelURL()},
-		})
-	}
-
-	if config.TosURL() != "" {
-		inlineKeyboard = append(inlineKeyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "tos_button"), URL: config.TosURL()},
-		})
-	}
-	return inlineKeyboard
-}
-
-func (h Handler) TrialCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	if config.TrialDays() == 0 {
-		return
-	}
-	c, err := h.customerRepository.FindByTelegramId(ctx, update.CallbackQuery.From.ID)
-	if err != nil {
-		slog.Error("Error finding customer", err)
-		return
-	}
-	if c == nil {
-		slog.Error("customer not exist", "telegramId", utils.MaskHalfInt64(update.CallbackQuery.From.ID), "error", err)
-		return
-	}
-	if c.SubscriptionLink != nil {
-		return
-	}
-	callback := update.CallbackQuery.Message.Message
-	langCode := update.CallbackQuery.From.LanguageCode
-	_, err = b.EditMessageText(ctx, &bot.EditMessageTextParams{
-		ChatID:    callback.Chat.ID,
-		MessageID: callback.ID,
-		Text:      h.translation.GetText(langCode, "trial_text"),
-		ParseMode: models.ParseModeHTML,
-		ReplyMarkup: models.InlineKeyboardMarkup{
-			InlineKeyboard: [][]models.InlineKeyboardButton{
-				{{Text: h.translation.GetText(langCode, "activate_trial_button"), CallbackData: CallbackActivateTrial}},
-				{{Text: h.translation.GetText(langCode, "back_button"), CallbackData: CallbackStart}},
-			},
-		},
-	})
-	if err != nil {
-		slog.Error("Error sending /trial message", err)
-	}
-}
-
-func (h Handler) ActivateTrialCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	if config.TrialDays() == 0 {
-		return
-	}
-	c, err := h.customerRepository.FindByTelegramId(ctx, update.CallbackQuery.From.ID)
-	if err != nil {
-		slog.Error("Error finding customer", err)
-		return
-	}
-	if c == nil {
-		slog.Error("customer not exist", "telegramId", utils.MaskHalfInt64(update.CallbackQuery.From.ID), "error", err)
-		return
-	}
-	if c.SubscriptionLink != nil {
-		return
-	}
-	callback := update.CallbackQuery.Message.Message
-	ctxWithUsername := context.WithValue(ctx, "username", update.CallbackQuery.From.Username)
-	_, err = h.paymentService.ActivateTrial(ctxWithUsername, update.CallbackQuery.From.ID)
-	langCode := update.CallbackQuery.From.LanguageCode
-	_, err = b.EditMessageText(ctx, &bot.EditMessageTextParams{
-		ChatID:      callback.Chat.ID,
-		MessageID:   callback.ID,
-		Text:        h.translation.GetText(langCode, "trial_activated"),
-		ParseMode:   models.ParseModeHTML,
-		ReplyMarkup: models.InlineKeyboardMarkup{InlineKeyboard: h.createConnectKeyboard(langCode)},
-	})
-	if err != nil {
-		slog.Error("Error sending /trial message", err)
-	}
-}
-
-func (h Handler) createConnectKeyboard(lang string) [][]models.InlineKeyboardButton {
-	var inlineCustomerKeyboard [][]models.InlineKeyboardButton
-	inlineCustomerKeyboard = append(inlineCustomerKeyboard, h.resolveConnectButton(lang))
-
-	inlineCustomerKeyboard = append(inlineCustomerKeyboard, []models.InlineKeyboardButton{
-		{Text: h.translation.GetText(lang, "back_button"), CallbackData: CallbackStart},
-	})
-	return inlineCustomerKeyboard
-}
-
-func (h Handler) BuyCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	callback := update.CallbackQuery.Message.Message
-	langCode := update.CallbackQuery.From.LanguageCode
-
-	var priceButtons []models.InlineKeyboardButton
-
-	if config.Price1() > 0 {
-		priceButtons = append(priceButtons, models.InlineKeyboardButton{
-			Text:         h.translation.GetText(langCode, "month_1"),
-			CallbackData: fmt.Sprintf("%s?month=%d&amount=%d", CallbackSell, 1, config.Price1()),
-		})
-	}
-
-	if config.Price3() > 0 {
-		priceButtons = append(priceButtons, models.InlineKeyboardButton{
-			Text:         h.translation.GetText(langCode, "month_3"),
-			CallbackData: fmt.Sprintf("%s?month=%d&amount=%d", CallbackSell, 3, config.Price3()),
-		})
-	}
-
-	if config.Price6() > 0 {
-		priceButtons = append(priceButtons, models.InlineKeyboardButton{
-			Text:         h.translation.GetText(langCode, "month_6"),
-			CallbackData: fmt.Sprintf("%s?month=%d&amount=%d", CallbackSell, 6, config.Price6()),
-		})
-	}
-
-	if config.Price12() > 0 {
-		priceButtons = append(priceButtons, models.InlineKeyboardButton{
-			Text:         h.translation.GetText(langCode, "month_12"),
-			CallbackData: fmt.Sprintf("%s?month=%d&amount=%d", CallbackSell, 12, config.Price12()),
-		})
-	}
-
-	keyboard := [][]models.InlineKeyboardButton{}
-
-	if len(priceButtons) == 4 {
-		keyboard = append(keyboard, priceButtons[:2])
-		keyboard = append(keyboard, priceButtons[2:])
-	} else if len(priceButtons) > 0 {
-		keyboard = append(keyboard, priceButtons)
-	}
-
-	keyboard = append(keyboard, []models.InlineKeyboardButton{
-		{Text: h.translation.GetText(langCode, "back_button"), CallbackData: CallbackStart},
-	})
-
-	_, err := b.EditMessageText(ctx, &bot.EditMessageTextParams{
-		ChatID:    callback.Chat.ID,
-		MessageID: callback.ID,
-		ParseMode: models.ParseModeHTML,
-		ReplyMarkup: models.InlineKeyboardMarkup{
-			InlineKeyboard: keyboard,
-		},
-		Text: h.translation.GetText(langCode, "pricing_info"),
-	})
-
-	if err != nil {
-		slog.Error("Error sending buy message", err)
-	}
-}
-func (h Handler) SellCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	callback := update.CallbackQuery.Message.Message
-	callbackQuery := parseCallbackData(update.CallbackQuery.Data)
-	langCode := update.CallbackQuery.From.LanguageCode
-	month := callbackQuery["month"]
-	amount := callbackQuery["amount"]
-
-	var keyboard [][]models.InlineKeyboardButton
-
-	if config.IsCryptoPayEnabled() {
-		keyboard = append(keyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "crypto_button"), CallbackData: fmt.Sprintf("%s?month=%s&invoiceType=%s&amount=%s", CallbackPayment, month, database.InvoiceTypeCrypto, amount)},
-		})
-	}
-
-	if config.IsYookasaEnabled() {
-		keyboard = append(keyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "card_button"), CallbackData: fmt.Sprintf("%s?month=%s&invoiceType=%s&amount=%s", CallbackPayment, month, database.InvoiceTypeYookasa, amount)},
-		})
-	}
-
-	if config.IsTelegramStarsEnabled() {
-		keyboard = append(keyboard, []models.InlineKeyboardButton{
-			{Text: h.translation.GetText(langCode, "stars_button"), CallbackData: fmt.Sprintf("%s?month=%s&invoiceType=%s&amount=%s", CallbackPayment, month, database.InvoiceTypeTelegram, amount)},
-		})
-	}
-
-	keyboard = append(keyboard, []models.InlineKeyboardButton{
-		{Text: h.translation.GetText(langCode, "back_button"), CallbackData: CallbackBuy},
-	})
-
-	_, err := b.EditMessageReplyMarkup(ctx, &bot.EditMessageReplyMarkupParams{
-		ChatID:    callback.Chat.ID,
-		MessageID: callback.ID,
-		ReplyMarkup: models.InlineKeyboardMarkup{
-			InlineKeyboard: keyboard,
-		},
-	})
-
-	if err != nil {
-		slog.Error("Error sending sell message", err)
-	}
-}
-
-func (h Handler) PaymentCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	callback := update.CallbackQuery.Message.Message
-	callbackQuery := parseCallbackData(update.CallbackQuery.Data)
-	month, err := strconv.Atoi(callbackQuery["month"])
-	if err != nil {
-		slog.Error("Error getting month from query", err)
-		return
-	}
-
-	price, err := strconv.Atoi(callbackQuery["amount"])
-	if err != nil {
-		slog.Error("Error getting price from query", err)
-		return
-	}
-
-	invoiceType := database.InvoiceType(callbackQuery["invoiceType"])
-
-	ctx, cancel := context.WithTimeout(context.Background(), time.Second*10)
-	defer cancel()
-	customer, err := h.customerRepository.FindByTelegramId(ctx, callback.Chat.ID)
-	if err != nil {
-		slog.Error("Error finding customer", err)
-		return
-	}
-	if customer == nil {
-		slog.Error("customer not exist", "chatID", callback.Chat.ID, "error", err)
-		return
-	}
-
-	ctxWithUsername := context.WithValue(ctx, "username", update.CallbackQuery.From.Username)
-	paymentURL, purchaseId, err := h.paymentService.CreatePurchase(ctxWithUsername, price, month, customer, invoiceType)
-	if err != nil {
-		slog.Error("Error creating payment", err)
-		return
-	}
-
-	langCode := update.CallbackQuery.From.LanguageCode
-
-	message, err := b.EditMessageReplyMarkup(ctx, &bot.EditMessageReplyMarkupParams{
-		ChatID:    callback.Chat.ID,
-		MessageID: callback.ID,
-		ReplyMarkup: models.InlineKeyboardMarkup{
-			InlineKeyboard: [][]models.InlineKeyboardButton{
-				{
-					{Text: h.translation.GetText(langCode, "pay_button"), URL: paymentURL},
-					{Text: h.translation.GetText(langCode, "back_button"), CallbackData: fmt.Sprintf("%s?month=%d&amount=%d", CallbackSell, month, price)},
-				},
-			},
-		},
-	})
-	if err != nil {
-		slog.Error("Error updating sell message", err)
-		return
-	}
-	h.cache.Set(purchaseId, message.ID)
-}
-
-func (h Handler) ConnectCommandHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	customer, err := h.customerRepository.FindByTelegramId(ctx, update.Message.Chat.ID)
-	if err != nil {
-		slog.Error("Error finding customer", err)
-		return
-	}
-	if customer == nil {
-		slog.Error("customer not exist", "telegramId", utils.MaskHalfInt64(update.Message.Chat.ID), "error", err)
-		return
-	}
-
-	langCode := update.Message.From.LanguageCode
-
-	isDisabled := true
-	_, err = b.SendMessage(ctx, &bot.SendMessageParams{
-		ChatID:    update.Message.Chat.ID,
-		Text:      buildConnectText(customer, langCode),
-		ParseMode: models.ParseModeHTML,
-		LinkPreviewOptions: &models.LinkPreviewOptions{
-			IsDisabled: &isDisabled,
-		},
-		ReplyMarkup: models.InlineKeyboardMarkup{
-			InlineKeyboard: [][]models.InlineKeyboardButton{
-				{{Text: h.translation.GetText(langCode, "back_button"), CallbackData: CallbackStart}},
-			},
-		},
-	})
-
-	if err != nil {
-		slog.Error("Error sending connect message", err)
-	}
-}
-
-func (h Handler) ConnectCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	callback := update.CallbackQuery.Message.Message
-
-	customer, err := h.customerRepository.FindByTelegramId(ctx, callback.Chat.ID)
-	if err != nil {
-		slog.Error("Error finding customer", err)
-		return
-	}
-	if customer == nil {
-		slog.Error("customer not exist", "telegramId", utils.MaskHalfInt64(callback.Chat.ID), "error", err)
-		return
-	}
-
-	langCode := update.CallbackQuery.From.LanguageCode
-
-	isDisabled := true
-	_, err = b.EditMessageText(ctx, &bot.EditMessageTextParams{
-		ChatID:    callback.Chat.ID,
-		MessageID: callback.ID,
-		ParseMode: models.ParseModeHTML,
-		Text:      buildConnectText(customer, langCode),
-		LinkPreviewOptions: &models.LinkPreviewOptions{
-			IsDisabled: &isDisabled,
-		},
-		ReplyMarkup: models.InlineKeyboardMarkup{
-			InlineKeyboard: [][]models.InlineKeyboardButton{
-				{{Text: h.translation.GetText(langCode, "back_button"), CallbackData: CallbackStart}},
-			},
-		},
-	})
-
-	if err != nil {
-		slog.Error("Error sending connect message", err)
-	}
-}
-
-func (h Handler) PreCheckoutCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	_, err := b.AnswerPreCheckoutQuery(ctx, &bot.AnswerPreCheckoutQueryParams{
-		PreCheckoutQueryID: update.PreCheckoutQuery.ID,
-		OK:                 true,
-	})
-	if err != nil {
-		slog.Error("Error sending answer pre checkout query", err)
-	}
-}
-
-func (h Handler) SuccessPaymentHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	payload := strings.Split(update.Message.SuccessfulPayment.InvoicePayload, "&")
-	purchaseId, err := strconv.Atoi(payload[0])
-	username := payload[1]
-	if err != nil {
-		slog.Error("Error parsing purchase id", err)
-		return
-	}
-
-	ctxWithUsername := context.WithValue(ctx, "username", username)
-	err = h.paymentService.ProcessPurchaseById(ctxWithUsername, int64(purchaseId))
-	if err != nil {
-		slog.Error("Error processing purchase", err)
-	}
-
-}
-
-func (h Handler) SyncUsersCommandHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
-	h.syncService.Sync()
-	_, err := b.SendMessage(ctx, &bot.SendMessageParams{
-		ChatID: update.Message.Chat.ID,
-		Text:   "Users synced",
-	})
-	if err != nil {
-		slog.Error("Error sending sync message", err)
-	}
-}
-
-func buildConnectText(customer *database.Customer, langCode string) string {
-	var info strings.Builder
-
-	tm := translation.GetInstance()
-
-	if customer.ExpireAt != nil {
-		currentTime := time.Now()
-
-		if currentTime.Before(*customer.ExpireAt) {
-			formattedDate := customer.ExpireAt.Format("02.01.2006 15:04")
-
-			subscriptionActiveText := tm.GetText(langCode, "subscription_active")
-			info.WriteString(fmt.Sprintf(subscriptionActiveText, formattedDate))
-
-			if customer.SubscriptionLink != nil && *customer.SubscriptionLink != "" {
-				subscriptionLinkText := tm.GetText(langCode, "subscription_link")
-				info.WriteString(fmt.Sprintf(subscriptionLinkText, *customer.SubscriptionLink))
-			}
-		} else {
-			noSubscriptionText := tm.GetText(langCode, "no_subscription")
-			info.WriteString(noSubscriptionText)
-		}
-	} else {
-		noSubscriptionText := tm.GetText(langCode, "no_subscription")
-		info.WriteString(noSubscriptionText)
-	}
-
-	return info.String()
-}
-
-func parseCallbackData(data string) map[string]string {
-	result := make(map[string]string)
-
-	parts := strings.Split(data, "?")
-	if len(parts) < 2 {
-		return result
-	}
-
-	params := strings.Split(parts[1], "&")
-	for _, param := range params {
-		kv := strings.SplitN(param, "=", 2)
-		if len(kv) == 2 {
-			result[kv[0]] = kv[1]
-		}
-	}
-
-	return result
 }
