+++ internal/handler/sync.go
@@ -1,0 +1,20 @@
+package handler
+
+import (
+	"context"
+
+	"github.com/go-telegram/bot"
+	"github.com/go-telegram/bot/models"
+	"log/slog"
+)
+
+func (h Handler) SyncUsersCommandHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
+	h.syncService.Sync()
+	_, err := b.SendMessage(ctx, &bot.SendMessageParams{
+		ChatID: update.Message.Chat.ID,
+		Text:   "Users synced",
+	})
+	if err != nil {
+		slog.Error("Error sending sync message", err)
+	}
+}
