+++ internal/handler/trial.go
@@ -1,0 +1,87 @@
+package handler
+
+import (
+	"context"
+
+	"github.com/go-telegram/bot"
+	"github.com/go-telegram/bot/models"
+	"log/slog"
+
+	"remnawave-tg-shop-bot/internal/config"
+	"remnawave-tg-shop-bot/utils"
+)
+
+func (h Handler) TrialCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
+	if config.TrialDays() == 0 {
+		return
+	}
+	c, err := h.customerRepository.FindByTelegramId(ctx, update.CallbackQuery.From.ID)
+	if err != nil {
+		slog.Error("Error finding customer", err)
+		return
+	}
+	if c == nil {
+		slog.Error("customer not exist", "telegramId", utils.MaskHalfInt64(update.CallbackQuery.From.ID), "error", err)
+		return
+	}
+	if c.SubscriptionLink != nil {
+		return
+	}
+	callback := update.CallbackQuery.Message.Message
+	langCode := update.CallbackQuery.From.LanguageCode
+	_, err = b.EditMessageText(ctx, &bot.EditMessageTextParams{
+		ChatID:    callback.Chat.ID,
+		MessageID: callback.ID,
+		Text:      h.translation.GetText(langCode, "trial_text"),
+		ParseMode: models.ParseModeHTML,
+		ReplyMarkup: models.InlineKeyboardMarkup{InlineKeyboard: [][]models.InlineKeyboardButton{
+			{{Text: h.translation.GetText(langCode, "activate_trial_button"), CallbackData: CallbackActivateTrial}},
+			{{Text: h.translation.GetText(langCode, "back_button"), CallbackData: CallbackStart}},
+		}},
+	})
+	if err != nil {
+		slog.Error("Error sending /trial message", err)
+	}
+}
+
+func (h Handler) ActivateTrialCallbackHandler(ctx context.Context, b *bot.Bot, update *models.Update) {
+	if config.TrialDays() == 0 {
+		return
+	}
+	c, err := h.customerRepository.FindByTelegramId(ctx, update.CallbackQuery.From.ID)
+	if err != nil {
+		slog.Error("Error finding customer", err)
+		return
+	}
+	if c == nil {
+		slog.Error("customer not exist", "telegramId", utils.MaskHalfInt64(update.CallbackQuery.From.ID), "error", err)
+		return
+	}
+	if c.SubscriptionLink != nil {
+		return
+	}
+	callback := update.CallbackQuery.Message.Message
+	ctxWithUsername := context.WithValue(ctx, "username", update.CallbackQuery.From.Username)
+	_, err = h.paymentService.ActivateTrial(ctxWithUsername, update.CallbackQuery.From.ID)
+	langCode := update.CallbackQuery.From.LanguageCode
+	_, err = b.EditMessageText(ctx, &bot.EditMessageTextParams{
+		ChatID:      callback.Chat.ID,
+		MessageID:   callback.ID,
+		Text:        h.translation.GetText(langCode, "trial_activated"),
+		ParseMode:   models.ParseModeHTML,
+		ReplyMarkup: models.InlineKeyboardMarkup{InlineKeyboard: h.createConnectKeyboard(langCode)},
+	})
+	if err != nil {
+		slog.Error("Error sending /trial message", err)
+	}
+}
+
+func (h Handler) createConnectKeyboard(lang string) [][]models.InlineKeyboardButton {
+	var inlineCustomerKeyboard [][]models.InlineKeyboardButton
+	inlineCustomerKeyboard = append(inlineCustomerKeyboard, h.resolveConnectButton(lang))
+
+	inlineCustomerKeyboard = append(inlineCustomerKeyboard, []models.InlineKeyboardButton{
+		{Text: h.translation.GetText(lang, "back_button"), CallbackData: CallbackStart},
+	})
+	return inlineCustomerKeyboard
+}
